% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{find putative song epochs}
\usage{
segment(spec, plots = FALSE, verbose = FALSE, freq_mad_threshold = 2,
  window_length = 3, threshold = 0.3)
}
\arguments{
\item{spec}{[matrix] Spectrogram bitmap, normalized to 1. Expected to have dim = (freq
x time).}

\item{plots, }{verbose [logical] flags for whether to show debugging plots and text}

\item{freq_mad_threshold}{frequency deviation threshold, in units of standard absolute median devs.}

\item{window_length}{moving average window length}

\item{threshold}{[0/1]}
}
\description{
How it works:
\enumerate{
  \item Find pixels that exceed some multiple of the absolute median deviation for each frequency. Specified by \code{freq_mad_threshold}.
  \item Smooth the time series of power deviations (window_length).
  \item Find time points when power deviates for > threshold number/proportion of frequencies.
}
}

